% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_gene_name_format.R
\name{check_ensembel}
\alias{check_ensembel}
\title{Check if Gene Names are ENSEMBL or SYMBOL}
\usage{
check_ensembel(input)
}
\arguments{
\item{input}{A matrix or Seurat object with gene names as rownames.}
}
\value{
A character string: either "ENSEMBEL" or "SYMBOL".
}
\description{
Determines whether the input matrix contains ENSEMBL IDs or gene symbols.
}
