% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_meta_cells.R
\name{compute_all_clusters_prob_matrix}
\alias{compute_all_clusters_prob_matrix}
\title{Aggregate Probabilities Across All Clusters}
\usage{
compute_all_clusters_prob_matrix(prob_data, cluster_labels, nCores = NULL)
}
\arguments{
\item{prob_data}{Probability matrix (cells x spatial spots).}

\item{cluster_labels}{A vector of cluster labels for each cell.}

\item{nCores}{Optional number of CPU cores to use.}
}
\value{
A sparse matrix (clusters x spatial spots).
}
\description{
Computes a new matrix of average probabilities for each cluster using parallelization.
}
