% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_meta_cells.R
\name{compute_single_cluster_prob}
\alias{compute_single_cluster_prob}
\title{Compute Probability Vector for a Single Cluster}
\usage{
compute_single_cluster_prob(prob_data, cluster_labels, cluster_id)
}
\arguments{
\item{prob_data}{Probability matrix (cells x spatial spots), rownames are cell IDs.}

\item{cluster_labels}{A data frame with columns 'cell' and 'cluster'.}

\item{cluster_id}{The cluster ID to compute the average for.}
}
\value{
A numeric vector of average probabilities for the selected cluster.
}
\description{
Computes the mean probabilities for all spatial regions given one cluster of scRNA-seq data.
}
