% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_gene_name_format.R
\name{convert_gene_names_scRNA}
\alias{convert_gene_names_scRNA}
\title{Convert Gene Names in scRNA-seq Data}
\usage{
convert_gene_names_scRNA(dat_scRNA, assay_name = "SCT")
}
\arguments{
\item{dat_scRNA}{A Seurat object to be modified in place.}

\item{assay_name}{Assay name (default: "SCT").}
}
\value{
A list with updated Seurat object and gene conversion data frame.
}
\description{
Converts between ENSEMBL IDs and gene symbols in the given Seurat object.
}
