% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_by_spot_prob.R
\name{create_correlation_matrix}
\alias{create_correlation_matrix}
\title{Cell by Spot Correlation Matrix}
\usage{
create_correlation_matrix(
  dat_scRNA,
  dat_visium,
  p_hvg = 2000,
  num_pca = 100,
  alpha = 0.05,
  nCores = NULL
)
}
\arguments{
\item{dat_scRNA}{A Seurat object or a matrix containing scRNA data.}

\item{dat_visium}{A Seurat object containing spatial transcriptomic data with a spatial assay.}

\item{p_hvg}{Number of highly variable genes to use from each data.}

\item{num_pca}{Number of principle components}

\item{alpha}{Level of significance of FDR control}

\item{nCores}{Number of cores to use for parallelization}
}
\value{
Returns the spot by cell correlation matrix, each column is one cell, and each row is a spatial spot
}
\description{
Computes sparse correlation matrix between the single cells and the spatial spots.
}
