% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_by_spot_prob.R
\name{create_probability_matrix}
\alias{create_probability_matrix}
\title{Cell by Spot Probability Matrix}
\usage{
create_probability_matrix(
  dat_scRNA,
  dat_visium,
  p_hvg = 2000,
  num_pca = 100,
  alpha = 0.05,
  nCores = NULL
)
}
\arguments{
\item{dat_scRNA}{A Seurat object or a matrix containing scRNA data.}

\item{dat_visium}{A Seurat object containing spatial transcriptomic data with a spatial assay.}

\item{p_hvg}{Number of highly variable genes to use from each data.}

\item{num_pca}{Number of principle components}

\item{alpha}{Level of significance of FDR control}

\item{nCores}{Number of cores to use for parallelization}
}
\value{
Returns the cell by spot probability matrix, each row corresponds to the spatial distribution of one cell.
}
\description{
Computes mapping probabilities of the single cells to the spatial spots.
}
