% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_meta_cells.R
\name{find_cluster_scRNA_HGC}
\alias{find_cluster_scRNA_HGC}
\title{Hierarchical Graph Clustering for scRNA-seq}
\usage{
find_cluster_scRNA_HGC(dat_scRNA, k, num_pca = 100, p_hvg = 2000)
}
\arguments{
\item{dat_scRNA}{Seurat object with SCT assay.}

\item{k}{Number of clusters to cut from hierarchical tree.}

\item{num_pca}{Number of principal components.}

\item{p_hvg}{Number of highly variable genes.}
}
\value{
A data frame with cell names and assigned cluster IDs.
}
\description{
Applies PCA and hierarchical clustering (HGC) to scRNA-seq data using Seurat and ClusteringTree.
}
