% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_seurat_data.R
\name{transform_scRNA_data}
\alias{transform_scRNA_data}
\title{SCTransform Normalization for scRNA-seq Data}
\usage{
transform_scRNA_data(dat_scRNA, thres = NULL, metadata = NULL)
}
\arguments{
\item{dat_scRNA}{A Seurat object or raw gene expression matrix (genes x cells).}

\item{thres}{Threshold for minimum number of cells a gene must be expressed in. If NULL, it defaults to 0.5\% of total cells.}

\item{metadata}{Optional metadata to use if a matrix is provided.}
}
\value{
A normalized Seurat object with SCT assay.
}
\description{
Applies SCTransform normalization to scRNA-seq data, accepting either a Seurat object or a raw matrix.
}
