% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data.R
\name{transform_spatial_data}
\alias{transform_spatial_data}
\title{Transform Spatial Data}
\usage{
transform_spatial_data(dat_visium, thres = NULL)
}
\arguments{
\item{dat_visium}{A Seurat object containing spatial transcriptomic data with a spatial assay.}

\item{thres}{Optional threshold for gene filtering. If NULL, defaults to 0.5\% of the number of genes.}
}
\value{
A Seurat object with transformed data.
}
\description{
Transforms Spatial Transcriptomic data using SCTransform from the Seurat package.
}
