% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data.R
\name{check_ensembel}
\alias{check_ensembel}
\title{Check Gene Name Format}
\usage{
check_ensembel(input)
}
\arguments{
\item{input}{A Seurat object containing scRNA/spatial transcriptomic data .}
}
\value{
Returns the format of the gene names, ENSEMBEL or SYMBOL
}
\description{
Checks the gene name format in scRNA/spatial data
}
