% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_distribution.R
\name{compute_all_clusters_prob_matrix}
\alias{compute_all_clusters_prob_matrix}
\title{Avg. Spatial probability of all clusters/meta-cells}
\usage{
compute_all_clusters_prob_matrix(prob_data, cluster_labels, nCores = NULL)
}
\arguments{
\item{prob_data}{The cell by spot probability matrix}

\item{cluster_labels}{Assigned cluster labels of the cells. Length should be equal to the nrow(prob_data), and the i^th entry should correspond to the i^th row of prob_data}

\item{nCores}{Number of cores for parallelization}
}
\value{
The spatial probability matrix of the meta-cells. Each row corresponds to one such cluster.
}
\description{
Avg. Spatial probability of all clusters/meta-cells
}
