% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_distribution.R
\name{compute_single_cluster_prob}
\alias{compute_single_cluster_prob}
\title{Avg. Spatial probability of a cluster/meta-cell}
\usage{
compute_single_cluster_prob(prob_data, cluster_labels, cluster_id)
}
\arguments{
\item{prob_data}{The cell by spot probability matrix}

\item{cluster_labels}{Assigned cluster labels of the cells. Length should be equal to the nrow(prob_data), and the i^th entry should correspond to the i^th row of prob_data}

\item{cluster_id}{The input cluster id}
}
\value{
The avg. probability distribution of the selected cluster
}
\description{
Avg. Spatial probability of a cluster/meta-cell
}
