% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_spot_by_cell_prob_matrix.R
\name{create_probability_matrix}
\alias{create_probability_matrix}
\title{Create Cross-Modality Probability Matrix}
\usage{
create_probability_matrix(
  dat_scRNA,
  dat_visium,
  p_hvg = 2000,
  num_pca = 100,
  alpha = 0.05,
  nCores = NULL
)
}
\arguments{
\item{dat_scRNA}{Seurat object with scRNA-seq data.}

\item{dat_visium}{Seurat object with spatial data.}

\item{p_hvg}{Number of highly variable genes to use.}

\item{num_pca}{Number of principal components.}

\item{alpha}{FDR significance threshold.}

\item{nCores}{Optional number of CPU cores to use.}
}
\value{
A sparse probability matrix (spatial clusters x scRNA clusters).
}
\description{
Constructs a sparse probability matrix between scRNA and Visium spatial data using PCA projections and a parallel FDR filtering step.
}
