% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_by_spot_prob.R
\name{create_probability_matrix}
\alias{create_probability_matrix}
\title{Cell by Spot Probability Matrix}
\usage{
create_probability_matrix(cor_matrix)
}
\arguments{
\item{cor_matrix}{A sparse or dense matrix of spot by cell correlation.}
}
\value{
Returns the cell by spot probability matrix, each row corresponds to the spatial distribution of one cell.
}
\description{
Computes mapping probabilities of the single cells to the spatial spots from the correlation matrix.
}
