% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_distribution.R
\name{find_cluster_scRNA_HGC}
\alias{find_cluster_scRNA_HGC}
\title{Cell clustering using HGC}
\usage{
find_cluster_scRNA_HGC(dat_scRNA, k, num_pca = 100, p_hvg = 2000)
}
\arguments{
\item{dat_scRNA}{A Seurat object containing scRNA data, with SCT assay}

\item{k}{Number of desired cell clusters or meta-cells}

\item{num_pca}{Number of PCA}

\item{p_hvg}{Number of highly variable genes}
}
\value{
Returns the cluster labels of each cell
}
\description{
Divide the single cell data into required number of clusters/meta-cells
}
