% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data.R
\name{transform_scRNA_data}
\alias{transform_scRNA_data}
\title{Transform scRNA Data}
\usage{
transform_scRNA_data(dat_scRNA, thres = NULL, metadata = NULL)
}
\arguments{
\item{dat_scRNA}{A Seurat object or a matrix containing scRNA data.}

\item{thres}{Optional threshold for gene filtering. If NULL, defaults to 0.5\% of the number of genes.}

\item{metadata}{Optional metadata to include in the Seurat object.}
}
\value{
A Seurat object with transformed data.
}
\description{
Transforms scRNA data using SCTransform from the Seurat package.
}
