% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_seurat_data.R
\name{transform_spatial_data}
\alias{transform_spatial_data}
\title{SCTransform Normalization for Visium Spatial Transcriptomics Data}
\usage{
transform_spatial_data(dat_visium, thres = NULL)
}
\arguments{
\item{dat_visium}{A Seurat object containing a "Spatial" assay.}

\item{thres}{Optional threshold for minimum cell expression per gene.}
}
\value{
A Seurat object with normalized Spatial assay.
}
\description{
Applies SCTransform to the Spatial assay of a Seurat object from Visium data.
}
